refactor: 조금 더 적절한 위치로 파일 이동
