+++ src/hooks/modal/useAlert.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { create } from 'zustand';
 
 /**
